Create Hello World program and Makefile.
Currently, this will only work on macOS. A late stage goal will be to
add Linux support. I am not planning on ever supporting Windows.

Currently, we are using the C library for IO. Whether or not this
indirectly violates the branching condition is TBD.

The Hello World code was gotten from:

    https://cs.lmu.edu/~ray/notes/x86assembly/

Signed-off-by: Joshua Brot <jbrot@umich.edu>