We now output the character count!
Implemented a number printing routine, which is pretty complicated.
Right now it uses one branch, which I will replace eventually once a
proper solution is found.

I realized the previous solution was still an indirect branch. Ideally,
I will fix it with self-modifying code---but getting that work is highly
non-trivial.

In the print_number routine, we can avoid a loop by making the function
fixed precision---although that's inelegant.

Signed-off-by: Joshua Brot <jbrot@umich.edu>